requirements.txt
Flask==3.0.2
Flask-CORS==4.0.1
python-dotenv==1.0.1
groq==0.4.0